bot/main.py
import logging
from telegram import Update
from telegram.ext import Application, CommandHandler, ContextTypes
from config import settings
from scheduler import seo_scheduler

# Setup logging
logging.basicConfig(
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    level=logging.INFO
)

logger = logging.getLogger(__name__)

async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Kirim pesan selamat datang saat perintah /start dipanggil"""
    welcome_text = (
        "🚀 Selamat datang di Robooster SEO Bot!\n\n"
        "Perintah yang tersedia:\n"
        "📊 /analyze <url> - Analisis SEO dasar\n"
        "📥 /submit <url> - Submit ke Google Indexing\n"
        "📋 /report - Dapatkan laporan SEO mingguan\n"
        "⏰ /remindme - Atur pengingat update konten"
    )
    await update.message.reply_text(welcome_text)

async def analyze(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Analisis SEO dasar untuk URL yang diberikan"""
    if not context.args:
        await update.message.reply_text("❌ Mohon berikan URL untuk dianalisis.\nContoh: /analyze https://example.com")
        return
    
    url = context.args[0]
    await update.message.reply_text(f"🔍 Menganalisis {url}...\nHasil akan segera dikirim.")
    # TODO: Implementasi analisis SEO

async def submit(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Submit URL ke Google Indexing API"""
    if not context.args:
        await update.message.reply_text("❌ Mohon berikan URL untuk disubmit.\nContoh: /submit https://example.com")
        return
    
    url = context.args[0]
    await update.message.reply_text(f"📤 Mensubmit {url} ke Google Indexing...")
    # TODO: Implementasi submit ke Google Indexing API

async def report(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Kirim laporan SEO mingguan"""
    await update.message.reply_text("📊 Menyiapkan laporan SEO mingguan...")
    # TODO: Implementasi generate dan kirim laporan

async def remindme(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Atur pengingat untuk update konten"""
    # TODO: Implementasi sistem pengingat
    await update.message.reply_text(
        "⏰ Fitur pengingat akan segera tersedia.\n"
        "Anda akan dapat mengatur jadwal untuk update konten."
    )

def main():
    """Start the bot"""
    # Create application
    application = Application.builder().token(settings.BOT_TOKEN).build()

    # Add handlers
    application.add_handler(CommandHandler("start", start))
    application.add_handler(CommandHandler("analyze", analyze))
    application.add_handler(CommandHandler("submit", submit))
    application.add_handler(CommandHandler("report", report))
    application.add_handler(CommandHandler("remindme", remindme))

    # Start scheduler
    seo_scheduler.start()

    # Run the bot
    application.run_polling()

if __name__ == '__main__':
    main()