bot/requirements.txt
python-telegram-bot>=20.0
apscheduler>=3.10.1
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0